Remove cyclic dependency between webserver and webserver-test-support:  - move io.helidon.webserver.testsupport.SocketHttpClient.java to webserver/src/test/java/ since it's only used in webserver  - remove the use of LoggingTestUtils from webserver
